feat: pass command line opts through to browser
Opts (as parsed by optimist) are exposed as the first argument
to the start function. Any arguments after a single '--' are not
processed, and are passed through as `.clientArgs`.

Starting the server sets the initial options, and they are
updated each time a new run is explicitly triggered.
Runs triggered by e.g file modifications use the most recent options.